Added ping script to get RTT on API
ping.py is a script that will insert n rows of pings into a given csv
file for processing in Excel.  We can use this to get the RTT from
various points on campus at various congestion levels to any echo
server.   Requires websocket-client package.  Use bing.bat to run on
Windows or ping.sh to run in bash.  The batch and bash scripts will need
to be updated when the server goes live so that it's not pinging
localhost, but rather our API server